+++ terra_ai/commands/__init__.py
@@ -1,6 +1,6 @@
 """Command handlers for TerraAI."""
 
-from terra_ai.commands.admin import AdminCommands
+from terra_ai.commands.management import ManagementCommands
 from terra_ai.commands.user import UserCommands
 
-__all__ = ["AdminCommands", "UserCommands"]
+__all__ = ["ManagementCommands", "UserCommands"]